+++ lib/ui/global_widget/app_card_doctor.dart
@@ -1,0 +1,57 @@
+import 'package:admin_clinica_front/core/constants/app_const_svgs.dart';
+import 'package:admin_clinica_front/ui/global_widget/app_text_style.dart';
+import 'package:admin_clinica_front/ui/view_models/doctor_view/doctor_view_models.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/svg.dart';
+
+class AppCardDoctor extends StatelessWidget {
+  const AppCardDoctor({
+    super.key,
+    required this.doctor,
+    this.width,
+  });
+
+  final DoctorsViewModel doctor;
+  final double? width;
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      width: width,
+      child: Card(
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(15.0),
+        ),
+        child: Row(
+          children: [
+            SvgPicture.asset(
+              AppConstSvgs.doctor,
+              height: 60,
+            ),
+            Expanded(
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Padding(
+                    padding: const EdgeInsets.symmetric(vertical: 2.5, horizontal: 8),
+                    child: AppTextGlobal.labelLightText(text: doctor.nombres),
+                  ),
+                  Padding(
+                    padding: const EdgeInsets.symmetric(horizontal: 8.0),
+                    child: Text(
+                      doctor.apellidos,
+                      style: const TextStyle(
+                        fontSize: 16,
+                        color: Colors.grey,
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
